# pp-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.BlockedUserResponse:
    properties:
      blocked_username:
        type: string
      blocker_username:
        type: string
      created_at:
        type: string
      id:
        type: string
    type: object
...
  handler.FriendRequestResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      receiver_username:
        type: string
      sender_username:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
  handler.UserResponse:
    properties:
      created_at:
        type: string
      is_active:
        type: boolean
      last_online_at:
        type: string
      nickname:
        type: string
      profile_picture_url:
        type: string
      role:
        type: string
      status_message:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pitturu.com
    name: PittuRu Team
    url: https://pitturu.com
  description: PittuRu 게임 플랫폼 REST API 서버
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PittuRu API
  version: "1.0"
paths:
...
  /me/blocked-users:
    get:
      description: Retrieves a list of users blocked by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.BlockedUserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: List blocked users
      tags:
      - Friend
...
  /me/friend-requests/{request_id}:
    delete:
      consumes:
      - application/json
      description: Cancels a pending outgoing friend request.
      parameters:
      - description: Friend request ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Cancel friend request
      tags:
      - Friend
  /me/friend-requests/{request_id}/accept:
    put:
      consumes:
      - application/json
      description: Accepts a pending friend request.
      parameters:
      - description: Friend request ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Accept friend request
      tags:
      - Friend
  /me/friend-requests/{request_id}/decline:
    put:
      consumes:
      - application/json
      description: Declines a pending friend request.
      parameters:
      - description: Friend request ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Decline friend request
      tags:
      - Friend
  /me/friend-requests/incoming:
    get:
      description: Retrieves a list of pending friend requests received by the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.FriendRequestResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: List incoming friend requests
      tags:
      - Friend
  /me/friend-requests/outgoing:
    get:
      description: Retrieves a list of pending friend requests sent by the authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.FriendRequestResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: List outgoing friend requests
      tags:
      - Friend
  /me/friends:
    get:
      description: Retrieves a list of the authenticated user's friends.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: List friends
      tags:
      - Friend
  /me/friends/{username}:
    delete:
      consumes:
      - application/json
      description: Removes an existing friend from the user's friend list.
      parameters:
      - description: Friend username to remove
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Remove friend
      tags:
      - Friend
...
  /users/{username}/block:
    post:
      consumes:
      - application/json
      description: Blocks another user, preventing communication and hiding activities.
      parameters:
      - description: Username to block
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.BlockedUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Block user
      tags:
      - Friend
  /users/{username}/friend-request:
    post:
      consumes:
      - application/json
      description: Sends a friend request to another user.
      parameters:
      - description: Receiver username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.FriendRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Send friend request
      tags:
      - Friend
...
  /users/{username}/unblock:
    delete:
      consumes:
      - application/json
      description: Unblocks a previously blocked user.
      parameters:
      - description: Username to unblock
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Unblock user
      tags:
      - Friend
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"